refactor(math): fold matrix domains under matrices
Fold the matrix-analysis, symbolic-matrix, certified-SNF, combinatorial-
matrix, canonical-form, and rational-linear domains in as subpackages of
the surviving matrices owner. Operation IDs and schemas are unchanged.

- matrix_analysis -> matrices/analysis
- symbolic_matrix -> matrices/symbolic
- certified_snf -> matrices/certified_snf
- combinatorial_matrices -> matrices/combinatorial
- canonical_forms -> matrices/canonical_forms
- rational_linear -> matrices/rational_linear

rational_linear folds with the matrix family because its contracts consume
matrices.values.RationalMatrix (exact rational linear systems, not
optimization); its operation IDs remain linear.*.

Validation: make check green (ruff, mypy, 977 tests); import-contracts
and architecture pass; snapshot unchanged at 360.
