+++ src/jacobian/math/matrices/analysis/__init__.py
