+++ src/jacobian/math/matrices/analysis/_models.py
