+++ src/jacobian/math/matrices/canonical_forms/__init__.py
@@ -1,6 +1,6 @@
 """Exact canonical-form kernels and typed contracts over QQ."""
 
-from jacobian.math.canonical_forms._models import (
+from jacobian.math.matrices.canonical_forms._models import (
     InvariantFactorEntry,
     MinimalPolynomialResult,
     MonicPolynomial,
@@ -8,7 +8,7 @@ from jacobian.math.canonical_forms._models import (
     RationalCanonicalFormResult,
     SquareMatrixRequest,
 )
-from jacobian.math.canonical_forms.operations import (
+from jacobian.math.matrices.canonical_forms.operations import (
     characteristic_polynomial,
     invariant_factors,
     minimal_polynomial,
