+++ src/jacobian/math/matrices/canonical_forms/_models.py
