+++ src/jacobian/math/matrices/certified_snf/_tools.py
@@ -1,7 +1,7 @@
 """Transformation-certified Smith normal forms."""
 
 from jacobian.catalog.models import MathTools
-from jacobian.math.certified_snf._operations import CERTIFIED_SNF_OPERATIONS
+from jacobian.math.matrices.certified_snf._operations import CERTIFIED_SNF_OPERATIONS
 
 __all__ = ["TOOLS"]
