+++ src/jacobian/math/matrices/combinatorial/__init__.py
