+++ src/jacobian/math/matrices/combinatorial/_models.py
