+++ src/jacobian/math/matrices/rational_linear/__init__.py
