+++ src/jacobian/math/matrices/rational_linear/_models.py
