+++ src/jacobian/math/matrices/rational_linear/_tools.py
@@ -1,7 +1,7 @@
 """Domain-owned exact rational-linear operations."""
 
 from jacobian.catalog.models import MathTools
-from jacobian.math.rational_linear._operations import (
+from jacobian.math.matrices.rational_linear._operations import (
     rational_linear_operations as _build_tools,
 )
