+++ src/jacobian/math/matrices/symbolic/__init__.py
@@ -1,6 +1,6 @@
 """Supported exact symbolic matrix API over QQ(t_1, ..., t_n)."""
 
-from jacobian.math.symbolic_matrix.operations import (
+from jacobian.math.matrices.symbolic.operations import (
     symbolic_characteristic_polynomial,
     symbolic_determinant,
     symbolic_eigenvalues,
