+++ src/jacobian/math/matrices/symbolic/_models.py
