+++ src/jacobian/math/matrices/symbolic/_tools.py
@@ -5,14 +5,14 @@ from collections.abc import Callable
 from jacobian._models import StrictModel
 from jacobian.catalog._examples import example
 from jacobian.catalog.models import MathTool, OperationExample
-from jacobian.math.symbolic_matrix._models import (
+from jacobian.math.matrices.symbolic._models import (
     SymbolicCharacteristicPolynomialResult,
     SymbolicDeterminantResult,
     SymbolicEigenvaluesResult,
     SymbolicMatrixRequest,
     SymbolicRankResult,
 )
-from jacobian.math.symbolic_matrix._operations import (
+from jacobian.math.matrices.symbolic._operations import (
     compute_symbolic_characteristic_polynomial,
     compute_symbolic_determinant,
     compute_symbolic_eigenvalues,
