+++ src/jacobian/math/matrices/symbolic/operations.py
