+++ tests/math/matrices/test_canonical_forms.py
@@ -6,13 +6,16 @@ import pytest
 from pydantic import ValidationError
 
 from jacobian._exact import CanonicalRational
-from jacobian.math.canonical_forms import (
+from jacobian.math.matrices.canonical_forms import (
     invariant_factors,
     minimal_polynomial,
     primary_decomposition,
 )
-from jacobian.math.canonical_forms._models import MonicPolynomial, SquareMatrixRequest
-from jacobian.math.canonical_forms._operations import (
+from jacobian.math.matrices.canonical_forms._models import (
+    MonicPolynomial,
+    SquareMatrixRequest,
+)
+from jacobian.math.matrices.canonical_forms._operations import (
     compute_minimal_polynomial,
     compute_primary_decomposition,
     compute_rational_canonical_form,
